Update feature theme colors, add shapes.
